requirements.txt
python-telegram-bot==22.5
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
google-generativeai==0.3.0
newsapi==0.1.1
pytz==2023.3
psycopg2-binary==2.9.9
aiohttp==3.9.1

